16-Industry Use Case-Human Resource-AutoQuestGen/Streamlit/requirements.txt
spacy==2.3.5
streamlit==0.79.0
pke==1.8.1
prettytable==2.1.0
flashtext==2.7
allennlp==0.9.0
tensorflow==2.1.0
Werkzeug==1.0.1
requests==2.25.1
pandas==1.1.3
pytest==0.0.0
transformers==2.8.0
Flask==1.1.2
nltk==3.5
ipython==7.22.0
Pillow==8.2.0
protobuf==3.15.8
PyPDF2==1.26.0
pyspark==3.1.1
vectors==1.0.0
sense2vec==1.0.2
